Monkey's Audio 4.11 (January 20, 2013)
1. NEW: Switched from MMX to SSE assembly, making compression and decompression faster.
2. NEW: Switched to latest Microsoft compiler, making everything around 5% faster.
3. Changed: Removed support for old Monkey's Audio 3.92 and earlier files (it hasn't been possible to make these files for over ten years -- use an old copy of Monkey's Audio to convert these files to the latest format if you have any).
4. Changed: Fixed a few compiler warnings.
5. Changed: Removed dependence on third-party NASM assembly compiler in favor of intrinsics (which compile nicely with all major compilers).
6. Changed: Put all Monkey's Audio source code into a namespace APE so that linking into third-party projects should never have name collisions.
Changed: Switched from #ifndef/endif to #pragma once (now supported by any reasonable compiler).

+++ Source/MACDll/CEEncoder.cpp
@@ -3,8 +3,7 @@
 #include <stdio.h>
 #include "filters.h" 
 #include "resource.h"
-
-#include "maclib.h"
+#include "MACLib.h"
 #include "CharacterHelper.h"
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -76,7 +75,7 @@ __declspec(dllexport) HANDLE FAR PASCAL OpenFilterOutput(LPSTR lpstrFilename,lon
     *lpChunkSize = 16384 * wfeAudioFormat.nBlockAlign;
 
     CSmartPtr<wchar_t> spUTF16(CAPECharacterHelper::GetUTF16FromANSI(lpstrFilename), TRUE);
-    if (pAPEEncoder->pAPECompress->Start(spUTF16, &wfeAudioFormat, (long)(float)lSize*rate, nCompressLevel, NULL, CREATE_WAV_HEADER_ON_DECOMPRESSION) != 0)
+    if (pAPEEncoder->pAPECompress->Start(spUTF16, &wfeAudioFormat, (int) (double(lSize) * rate), nCompressLevel, NULL, CREATE_WAV_HEADER_ON_DECOMPRESSION) != 0)
     {
         SafeDeleteAPEEncoder(pAPEEncoder);
         return NULL;
